Chemical contains

Chemical contains

Usage

chemical_contains(
  word = NULL,
  API_key = NULL,
  Server = chemical_api_server,
  verbose = FALSE,
  top = NULL
)

Arguments

word

A character string of a chemical name or portion of a chemical name

API_key

The user-specific API key

Server

The root address for the API endpoint

verbose

A logical indicating if some “progress report” should be given.

top

The number of results to return if there are multiple results available

Value

A data.frame of chemicals and related values matching the query parameters

Author

Paul Kruse, Kristin Issacs

Examples

if (FALSE) { # has_ctx_key() & is.na(ctx_key() == "FAKE_KEY")
# Pull chemicals that contain substring
substring_chemicals <- chemical_contains(word = 'TXSID702018')
}